calls abstracted to database.ts and implementation.ts
