+++ src/scripts/FireAdmin.ts
@@ -31,7 +31,7 @@ var adminAuth = adminInstance.auth();
  * returns a promise that is resolved with the user authentication object
  */
 export function registerUser(userEmail: string): Promise<any> {
-  return getUserPrivilege().then((privilege) => {
+  return getUserPrivilege().then((privilege: string) => {
     if (["Owner", "Admin"].includes(privilege)) {
       return adminAuth
         .createUserWithEmailAndPassword(userEmail, "yadaDefault")
